+++ rust/src/prototype/time_of_day_speed.rs
@@ -89,10 +89,10 @@ impl TimeOfDaySpeeds {
         link: &Link,
         second_of_day: SecondOfDay,
         day_of_week: DayOfWeek, 
-    ) -> u32 {
+    ) -> usize {
         if let Some(profile_id) = link.week_profile_ids[day_of_week] {
             let modifier = self.get_modifier_by_second_of_day(profile_id, second_of_day);
-            (link.time_seconds() as f64 * (1.0/modifier)) as u32
+            (link.time_seconds() as f64 * (1.0/modifier)) as usize 
 
         } else {
             link.time_seconds()
